coffee-price-monitor/requirements.txt
pandas
requests
beautifulsoup4
selenium
sqlalchemy
psycopg2-binary
python-dotenv